[!] Bug #0000437 fixed: expert must have an option to choose the needed status translation schema for Away, BRB & N/A statuses.
git-svn-id: http://miranda.googlecode.com/svn/trunk@623 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/MSN/msn_ssl.cpp
@@ -105,7 +105,7 @@ static char* sttSslGet( char* parUrl, char* parChallenge )
 
 	int tUsesProxy = MSN_GetByte( "NLUseProxy", 0 );
 	if ( tUsesProxy ) {
-		BYTE ptype = MSN_GetByte( "NLProxyType", 0 );
+		DWORD ptype = MSN_GetByte( "NLProxyType", 0 );
 		if ( !MSN_GetByte( "UseIeProxy", 0 ) && ( ptype == PROXYTYPE_HTTP || ptype == PROXYTYPE_HTTPS )) {
 			char szProxy[ 100 ];
 			if ( MSN_GetStaticString( "NLProxyServer", NULL, szProxy, sizeof szProxy )) {
